requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
peewee==3.14.4
peewee-migrate==1.4.0
PyYAML==5.4.1
six==1.16.0
Werkzeug==2.0.1
